Update default.conf
Non-transparent default color makes --background parameter not working. Basically the background image is loaded, drawon on the screen, but then the solid color is filled on top of it, so it looks like user did not specify background image at all. Making the default background color transparent makes it work as expected (at least the background image is still visible so user knows it is loaded). On a similar note, the same color should be adjusted in .c file. Will do in a separate patch.