Add option to override poll timeout while reading stdin
This change allows the user to specify how long xlunch will wait for
data on stdin before deciding that it is invalid. This might be
useful when a program is generating output by processing data
on-the-fly, and does not begin writing the data stream within
10 milliseconds (which is the timeout xlunch uses prior to this
change. It will also benefit older systems, especially those with
hard drives.